Authorization implemented

+++ src/api/notes/handler.js
@@ -17,8 +17,12 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { title = 'untitled', body, tags } = request.payload
+      const { id: credentialsId } = request.auth.credentials
 
-      const noteId = await this._service.addNote({ title, body, tags })
+      const noteId = await this._service.addNote({
+        title, body, tags, owner: credentialsId,
+      })
+
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil ditambahkan',
@@ -49,8 +53,10 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes()
+  async getNotesHandler(request) {
+    const { id: credentialsId } = request.auth.credentials
+
+    const notes = await this._service.getNotes(credentialsId)
     return {
       status: 'success',
       data: {
@@ -62,6 +68,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params
+      const { id: credentialsId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialsId)
       const note = await this._service.getNotesById(id)
       return {
         status: 'success',
@@ -94,6 +103,9 @@ class NotesHandler {
     try {
       this._validator.validateNotePayload(request.payload)
       const { id } = request.params
+      const { id: credentialsId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialsId)
       await this._service.editNoteById(id, request.payload)
 
       return {
@@ -124,6 +136,9 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params
+      const { id: credentialsId } = request.auth.credentials
+
+      await this._service.verifyNoteOwner(id, credentialsId)
 
       await this._service.deleteNoteById(id)
